Move EditorTypeTrait to Editor package
